W-073 - move enums to other package
