+++ backend/plortal/src/main/java/com/plortal/plortal/model/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.plortal.plortal.model;
+package com.plortal.plortal.model.enums;
 
 public enum UserRole {
     ADMIN,
