+++ backend/plortal/src/main/java/com/plortal/plortal/repository/LocationRepository.java
@@ -1,7 +1,7 @@
 package com.plortal.plortal.repository;
 
 import com.plortal.plortal.model.Location;
-import com.plortal.plortal.model.LocationType;
+import com.plortal.plortal.model.enums.LocationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
